+++ grid2op/Observation/CompleteObservation.py
@@ -105,11 +105,13 @@ class CompleteObservation(BaseObservation):
         "time_before_cooldown_line", "time_before_cooldown_sub",
         "time_next_maintenance", "duration_next_maintenance",
         "target_dispatch", "actual_dispatch",
-        # TODO: backward compatibility
         "storage_charge", "storage_power_target", "storage_power",
         "gen_p_before_curtail", "curtailment", "curtailment_limit"
     ]
-    attr_list_json = ["_shunt_p", "_shunt_q", "_shunt_v", "_shunt_bus"]
+    attr_list_json = ["_shunt_p", "_shunt_q", "_shunt_v", "_shunt_bus",
+                      "_thermal_limit",
+                      "support_theta",
+                      "theta_or", "theta_ex", "load_theta", "gen_theta", "storage_theta"]
     attr_list_set = set(attr_list_vect)
 
     def __init__(self,
@@ -199,3 +201,8 @@ class CompleteObservation(BaseObservation):
             self.curtailment[:] = 0.
             self.gen_p_before_curtail[:] = self.gen_p
             self.curtailment_limit[:] = 1.0
+
+        if env.backend.can_output_theta:
+            self.support_theta = True  # backend supports the computation of theta
+            self.theta_or[:], self.theta_ex[:], self.load_theta[:], self.gen_theta[:], self.storage_theta[:] = \
+                env.backend.get_theta()